feat: refactor naming of splash
